ylamsiah

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
  */
 void print_number(int n)
 {
-	long	nbr;
+	unsigned int	nbr;
 
 	nbr = n;
 	if (nbr >= 0 && nbr <= 9)
